added TravLockList, Menu
Added TravLockList, a list that can be traversed and locked.

Added LogicList, a TravLockList that has a second list containing game logic to be called by the LogicHandler, similar to Options which are being phased out.

Added MenuSpecs, a dataclass containing the nuts and bolts needed to build a Menu

Added Menu, a TravLockList  that can blit options.

Beginning transition from class based logic (AbstractLogic) to subroutines/functions.